Update bit sizes

+++ CSES/Introductory_Problems/AppleDivision.cc
@@ -1,10 +1,11 @@
 #include <climits>
+#include <cstdint>
 #include <iostream>
 #include <vector>
 
-using i64 = long long;
-using u64 = unsigned long long;
-using u32 = unsigned;
+using i64 = int64_t;
+using u64 = uint64_t;
+using u32 = uint32_t;
 using u128 = unsigned __int128;
 
 int main() {
